requirements.txt
dbt-core==1.10.19
dbt-sqlite==1.10.0
dlt==1.21.0
openpyxl==3.1.5
python-dotenv==1.2.1
